Move PyISPyBRestClient into abstract directory
